refactor(rest): move canonicalize endpoint from /system to /content
Move the canonicalize endpoint to POST /content/canonicalize as suggested
in review — it's a stateless content utility that fits alongside the
existing /content/references endpoint.

- Move implementation from SystemResourceImpl to ContentResourceImpl
- Move tests from SystemResourceTest to ContentResourceTest
- Restore SystemResourceImpl to its original state
- Regenerate Go SDK
