+++ app/src/main/java/io/apicurio/registry/rest/v3/impl/SystemResourceImpl.java
@@ -6,8 +6,6 @@ import io.apicurio.registry.auth.AuthConfig;
 import io.apicurio.registry.auth.Authorized;
 import io.apicurio.registry.auth.AuthorizedLevel;
 import io.apicurio.registry.auth.AuthorizedStyle;
-import io.apicurio.registry.content.ContentHandle;
-import io.apicurio.registry.content.TypedContent;
 import io.apicurio.registry.core.System;
 import io.apicurio.registry.limits.RegistryLimitsConfiguration;
 import io.apicurio.registry.logging.Logged;
@@ -20,18 +18,11 @@ import io.apicurio.registry.rest.v3.beans.UserInterfaceConfigAuth;
 import io.apicurio.registry.rest.v3.beans.UserInterfaceConfigFeatures;
 import io.apicurio.registry.rest.v3.beans.UserInterfaceConfigUi;
 import io.apicurio.registry.storage.impl.search.ElasticsearchSearchConfig;
-import io.apicurio.registry.storage.impl.sql.RegistryStorageContentUtils;
-import io.apicurio.registry.types.provider.ArtifactTypeUtilProviderFactory;
 import io.apicurio.registry.ui.UserInterfaceConfigProperties;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.interceptor.Interceptors;
-import jakarta.ws.rs.BadRequestException;
-import jakarta.ws.rs.core.Context;
-import jakarta.ws.rs.core.HttpHeaders;
-import jakarta.ws.rs.core.Response;
 
-import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -58,15 +49,6 @@ public class SystemResourceImpl implements SystemResource {
     @Inject
     ElasticsearchSearchConfig esSearchConfig;
 
-    @Inject
-    RegistryStorageContentUtils contentUtils;
-
-    @Inject
-    ArtifactTypeUtilProviderFactory factory;
-
-    @Context
-    HttpHeaders httpHeaders;
-
     /**
      * @see io.apicurio.registry.rest.v3.SystemResource#getSystemInfo()
      */
@@ -79,23 +61,6 @@ public class SystemResourceImpl implements SystemResource {
         info.setVersion(system.getVersion());
         info.setBuiltOn(system.getDate());
         return info;
-    }
-
-    @Override
-    @Authorized(style = AuthorizedStyle.None, level = AuthorizedLevel.Read)
-    public Response canonicalizeContent(String artifactType, InputStream data) {
-        ContentHandle content = ContentHandle.create(data);
-        if (content.bytes().length == 0) {
-            throw new BadRequestException("Empty content is not allowed.");
-        }
-        if (!factory.getAllArtifactTypes().contains(artifactType)) {
-            throw new BadRequestException("Unknown artifact type: " + artifactType);
-        }
-        String ct = httpHeaders.getMediaType() != null ? httpHeaders.getMediaType().toString() : null;
-        TypedContent typedContent = TypedContent.create(content, ct);
-        TypedContent canonicalized = contentUtils.canonicalizeContent(artifactType, typedContent,
-                Map.of());
-        return Response.ok(canonicalized.getContent()).type(canonicalized.getContentType()).build();
     }
 
     /**
