+++ app/src/test/java/io/apicurio/registry/noprofile/rest/v3/ContentResourceTest.java
@@ -11,6 +11,8 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
+import static io.restassured.RestAssured.given;
+
 @QuarkusTest
 public class ContentResourceTest extends AbstractResourceTestBase {
 
@@ -163,5 +165,127 @@ public class ContentResourceTest extends AbstractResourceTestBase {
             Assertions.assertNull(ref.getArtifactId());
             Assertions.assertNull(ref.getVersion());
         }
+    }
+
+    @Test
+    public void testCanonicalizeJsonContent() {
+        String content = "{\"z\": 1, \"a\": 2, \"m\": 3}";
+        String expected = "{\"a\":2,\"m\":3,\"z\":1}";
+
+        String actual = given().when().contentType(CT_JSON)
+                .queryParam("artifactType", ArtifactType.JSON)
+                .body(content)
+                .post("/registry/v3/content/canonicalize")
+                .then().statusCode(200)
+                .extract().body().asString();
+        Assertions.assertEquals(expected, actual);
+    }
+
+    @Test
+    public void testCanonicalizeOpenApiContent() {
+        String content = "{\n  \"openapi\": \"3.0.2\",\n  \"info\": {\n    \"title\": \"Empty API\","
+                + "\n    \"version\": \"1.0.0\"\n  },\n  \"paths\": {\n    \"/\": {}\n  },"
+                + "\n  \"components\": {}\n}";
+        String expected = "{\"components\":{},\"info\":{\"title\":\"Empty API\",\"version\":\"1.0.0\"},"
+                + "\"openapi\":\"3.0.2\",\"paths\":{\"/\":{}}}";
+
+        String actual = given().when().contentType(CT_JSON)
+                .queryParam("artifactType", ArtifactType.OPENAPI)
+                .body(content)
+                .post("/registry/v3/content/canonicalize")
+                .then().statusCode(200)
+                .extract().body().asString();
+        Assertions.assertEquals(expected, actual);
+    }
+
+    @Test
+    public void testCanonicalizeAvroContent() {
+        String content = "{\n  \"type\": \"record\",\n  \"namespace\": \"com.example\","
+                + "\n  \"name\": \"FullName\",\n  \"fields\": [\n"
+                + "    { \"name\": \"first\", \"type\": \"string\" }\n  ]\n}";
+        String expected = "{\"type\":\"record\",\"name\":\"FullName\",\"namespace\":\"com.example\","
+                + "\"doc\":\"\",\"fields\":[{\"name\":\"first\",\"type\":\"string\",\"doc\":\"\"}]}";
+
+        String actual = given().when().contentType(CT_JSON)
+                .queryParam("artifactType", ArtifactType.AVRO)
+                .body(content)
+                .post("/registry/v3/content/canonicalize")
+                .then().statusCode(200)
+                .extract().body().asString();
+        Assertions.assertEquals(expected, actual);
+    }
+
+    @Test
+    public void testCanonicalizeEmptyBody() {
+        given().when().contentType(CT_JSON)
+                .queryParam("artifactType", ArtifactType.JSON)
+                .body("")
+                .post("/registry/v3/content/canonicalize")
+                .then().statusCode(400);
+    }
+
+    @Test
+    public void testCanonicalizeUnknownArtifactType() {
+        given().when().contentType(CT_JSON)
+                .queryParam("artifactType", "GARBAGE")
+                .body("{\"key\": \"value\"}")
+                .post("/registry/v3/content/canonicalize")
+                .then().statusCode(400);
+    }
+
+    @Test
+    public void testCanonicalizeStoredArtifactContent() throws Exception {
+        String groupId = "testCanonicalizeStoredContent";
+        String artifactId = "testCanonicalizeStoredContent/JsonSchema";
+        String content = "{\"z\": 1, \"a\": 2, \"m\": 3}";
+        String expected = "{\"a\":2,\"m\":3,\"z\":1}";
+
+        createArtifact(groupId, artifactId, ArtifactType.JSON, content, ContentTypes.APPLICATION_JSON);
+
+        String actual = given().when()
+                .pathParam("groupId", groupId)
+                .pathParam("artifactId", artifactId)
+                .queryParam("canonical", true)
+                .get("/registry/v3/groups/{groupId}/artifacts/{artifactId}/versions/branch=latest/content")
+                .then().statusCode(200)
+                .extract().body().asString();
+        Assertions.assertEquals(expected, actual);
+    }
+
+    @Test
+    public void testGetStoredContentWithoutCanonical() throws Exception {
+        String groupId = "testGetStoredContentWithoutCanonical";
+        String artifactId = "testGetStoredContentWithoutCanonical/JsonSchema";
+        String content = "{\"z\": 1, \"a\": 2, \"m\": 3}";
+
+        createArtifact(groupId, artifactId, ArtifactType.JSON, content, ContentTypes.APPLICATION_JSON);
+
+        String actual = given().when()
+                .pathParam("groupId", groupId)
+                .pathParam("artifactId", artifactId)
+                .get("/registry/v3/groups/{groupId}/artifacts/{artifactId}/versions/branch=latest/content")
+                .then().statusCode(200)
+                .extract().body().asString();
+        Assertions.assertEquals(content, actual);
+    }
+
+    @Test
+    public void testCanonicalizeWithReferencesPreserveAndCanonical() throws Exception {
+        String groupId = "testCanonicalizeWithReferences";
+        String artifactId = "testCanonicalizeWithReferences/JsonSchema";
+        String content = "{\"z\": 1, \"a\": 2, \"m\": 3}";
+        String expected = "{\"a\":2,\"m\":3,\"z\":1}";
+
+        createArtifact(groupId, artifactId, ArtifactType.JSON, content, ContentTypes.APPLICATION_JSON);
+
+        String actual = given().when()
+                .pathParam("groupId", groupId)
+                .pathParam("artifactId", artifactId)
+                .queryParam("references", "PRESERVE")
+                .queryParam("canonical", true)
+                .get("/registry/v3/groups/{groupId}/artifacts/{artifactId}/versions/branch=latest/content")
+                .then().statusCode(200)
+                .extract().body().asString();
+        Assertions.assertEquals(expected, actual);
     }
 }
