+++ go-sdk/pkg/registryclient-v3/content/canonicalize_request_builder.go
@@ -1,4 +1,4 @@
-package system
+package content
 
 import (
 	"context"
@@ -30,7 +30,7 @@ type CanonicalizeRequestBuilderPostRequestConfiguration struct {
 // NewCanonicalizeRequestBuilderInternal instantiates a new CanonicalizeRequestBuilder and sets the default values.
 func NewCanonicalizeRequestBuilderInternal(pathParameters map[string]string, requestAdapter i2ae4187f7daee263371cb1c977df639813ab50ffa529013b7437480d1ec0158f.RequestAdapter) *CanonicalizeRequestBuilder {
 	m := &CanonicalizeRequestBuilder{
-		BaseRequestBuilder: *i2ae4187f7daee263371cb1c977df639813ab50ffa529013b7437480d1ec0158f.NewBaseRequestBuilder(requestAdapter, "{+baseurl}/system/canonicalize?artifactType={artifactType}", pathParameters),
+		BaseRequestBuilder: *i2ae4187f7daee263371cb1c977df639813ab50ffa529013b7437480d1ec0158f.NewBaseRequestBuilder(requestAdapter, "{+baseurl}/content/canonicalize?artifactType={artifactType}", pathParameters),
 	}
 	return m
 }
