+++ go-sdk/pkg/registryclient-v3/kiota-lock.json
@@ -1,5 +1,5 @@
 {
-  "descriptionHash": "4DEB22E479CDBAB81A6ADE8CF99502944A5B4FDBD9354E939605C1772FC0C99563E083F1F10AF464C20BAA99C60BF19DE6330AD80C13848642836EF7B435C776",
+  "descriptionHash": "6731D4611B08547330EE5EC10746160FF7D4C83954CD3895196F23C98495333E69C621A57E9B01770A375F59286D96D8225A884515EB02AC2638055F669D5386",
   "descriptionLocation": "../../v3.json",
   "lockFileVersion": "1.0.0",
   "kiotaVersion": "1.23.0",
